+++ libunhv3/unhv3.cpp
@@ -4,10 +4,15 @@
 #include "bondchunkattr.hpp"
 #include "unhv3.hpp"
 
-Unhv3::Unhv3() :
-    pwd(QDir::currentPath())
+Unhv3::Unhv3()
 {
     status = Unhv3Status::NO_ERROR;
+    converter = new HdpConverter();
+}
+
+Unhv3::~Unhv3()
+{
+    delete converter;
 }
 
 /** 마지막으로 발생한 오류의 상세 내용을 확인한다.
@@ -16,6 +21,56 @@ Unhv3::Unhv3() :
 Unhv3Status Unhv3::getLastError()
 {
     return status;
+}
+
+bool Unhv3::testArchive()
+{
+    status = Unhv3Status::NOT_YET_IMPELEMENTED;
+    return false;
+}
+
+bool Unhv3::isBrokenArchive()
+{
+    status = Unhv3Status::NOT_YET_IMPELEMENTED;
+    return false;
+}
+
+QString Unhv3::filePathName()
+{
+    return file.fileName();
+}
+
+bool Unhv3::isOpened()
+{
+    return file.isOpen();
+}
+
+void Unhv3::clear()
+{
+    file.close();
+    delete converter;
+    converter = new HdpConverter();
+    HV30_ = BondChunkHeader();
+    VERS_ = 0;
+    FSIZ_ = 0;
+    HEAD_ = BondChunkHeader();
+    GUID_ = QUuid();
+    UUID_ = QUuid();
+    FTIM_ = QDateTime();
+    DIRE_ = 0;
+    ENCR_ = 0;
+    COPY_ = QString::null;
+    LINK_ = QString::null;
+    TITL_ = QString::null;
+    ISBN_ = QString::null;
+    WRTR_ = QString::null;
+    PUBL_ = QString::null;
+    DATE_ = QString::null;
+    COMT_ = QString::null;
+    MAKR_ = QString::null;
+    GENR_ = QString::null;
+    LIST_ = FileInfoList();
+    BODY_ = FileDataStorage();
 }
 
 /** 지정된 경로에 파일을 모두 푼다.
@@ -42,11 +97,12 @@ bool Unhv3::extractAllTo(
     }
 
     int max = getFileItemCount();
+    QDir pwd = QDir::currentPath();
     pwd.cd(savePath);
     for(int i = 0; i < max; i++) {
         const FileInfo *fileInfo = getFileItem(i);
         QString fileName = fileInfo->NAME();
-        extractOneTo(i, fileName);
+        extractOneAs(i, fileName);
     }
     pwd.cd(".");
 
@@ -63,7 +119,7 @@ int Unhv3::getFileItemCount()
 
 const FileInfo* Unhv3::getFileItem(
         int index
-        )
+        ) const
 {
     return LIST_.getFileItem(index);
 }
@@ -73,17 +129,26 @@ bool Unhv3::extractOneTo(
         const QString &savePath
         )
 {
+    QString filePathName = savePath + "/" + getFileItem(index)->NAME();
+    return extractOneAs(index, filePathName);
+}
+
+bool Unhv3::extractOneAs(
+        int index,
+        const QString &filePathName
+        )
+{
     bool success = true;
     uint pos = LIST_.getFileItem(index)->POS4();
     QByteArray *raw_data = BODY_.getFileData(pos)->raw_data();
 
     try {
-        HdpConverter::getInstance()->setData(*raw_data);
-        if ( HdpConverter::getInstance()->hasAlphaChannel() ) {
-            HdpConverter::getInstance()->saveToPng(savePath);
+        converter->setData(*raw_data);
+        if ( converter->hasAlphaChannel() ) {
+            converter->saveToPng(filePathName);
         }
         else {
-            HdpConverter::getInstance()->saveToJpeg(savePath);
+            converter->saveToJpeg(filePathName);
         }
     }
     catch (WMP_err &err) {
@@ -102,9 +167,6 @@ bool Unhv3::open(
         const QString &filepath ///< 파일 경로
         )
 {
-    if ( file.isOpen() ) {
-        file.close();
-    }
     file.setFileName(filepath);
     fileStream_.setDevice(&file);