+++ libunhv3/unhv3status.hpp
@@ -24,13 +24,14 @@ class Unhv3Status
 {
 public:
     enum Status {
-        NO_ERROR,           ///< 오류 없음
-        FILE_NOT_EXIST,     ///< 파일 존재하지 않음
-        CANT_READ_FILE,     ///< 파일을 읽을수 없음
-        NOT_HV3_FORMAT,     ///< hv3 포멧이 아님
-        CRC_ERROR,          ///< CRC 오류
-        SAVEPATH_NOT_EXIST, ///< 지정된 저장 경로가 존재하지 않음.
-        SAVEPATH_IS_NOT_DIR ///< 지정된 저장 경로는 폴더가 아닙니다.
+        NO_ERROR,             ///< 오류 없음
+        FILE_NOT_EXIST,       ///< 파일 존재하지 않음
+        CANT_READ_FILE,       ///< 파일을 읽을수 없음
+        NOT_HV3_FORMAT,       ///< hv3 포멧이 아님
+        CRC_ERROR,            ///< CRC 오류
+        SAVEPATH_NOT_EXIST,   ///< 지정된 저장 경로가 존재하지 않음.
+        SAVEPATH_IS_NOT_DIR,  ///< 지정된 저장 경로는 폴더가 아닙니다.
+        NOT_YET_IMPELEMENTED  ///< 아직 구현되지 않은 기능입니다.
     };
     Unhv3Status& operator=(const Status &status);
     QString getMessage();